# Improve Base64.urlsafe_encode64 performance