Creating this here as asked in https://github.com/ruby/ruby/pull/4802

Improves the method's performance when asked to remove padding.

str.delete!("=") iterates over the entire string looking for the equals
character, but we know that we will, at most, find two at the end of the
string.