@jcmfernandes 2021-09-02
> `str.chomp!("=") and str.chomp!("=")` or `str.chomp!("==") or str.chomp!("=")` will work too.

Beautiful! :heart_eyes: